+++ src/test/java/io/github/astrapi69/string/StringExtensionsTest.java
@@ -29,11 +29,11 @@ import static org.testng.AssertJUnit.assertTrue;
 import java.util.List;
 import java.util.Map;
 
-import io.github.astrapi69.collections.array.ArrayFactory;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import io.github.astrapi69.BaseTestCase;
+import io.github.astrapi69.collections.array.ArrayFactory;
 import io.github.astrapi69.collections.list.ListFactory;
 import io.github.astrapi69.collections.map.MapFactory;
 import io.github.astrapi69.test.objects.Person;
@@ -46,6 +46,24 @@ import io.github.astrapi69.test.objects.Person;
  */
 public class StringExtensionsTest extends BaseTestCase
 {
+
+	/**
+	 * Test method for test the method {@link StringExtensions#getUniqueCharacters(String)}
+	 */
+	@Test
+	public void testGetUniqueCharacters()
+	{
+		List<Character> actual;
+		List<Character> expected;
+		String testString;
+
+		testString = "top secret";
+		actual = StringExtensions.getUniqueCharacters(testString);
+		expected = ListFactory.newArrayList(Character.valueOf(' '), Character.valueOf('c'),
+			Character.valueOf('e'), Character.valueOf('o'), Character.valueOf('p'),
+			Character.valueOf('r'), Character.valueOf('s'), Character.valueOf('t'));
+		assertEquals(actual, expected);
+	}
 
 	/**
 	 * Test method for {@link StringExtensions#addSingleQuotationMarks(String)}.